Collapse core checks to fewer extensions
Combine `EqualityChecks`, `HasField`, `TypeChecks`, into `CoreChecks`.
Move `StringChecks` to a separate library, it will continue to get
bigger. Leave `BoolChecks` in the core check library since it is so
small.
