+++ pkgs/checks/lib/checks.dart
@@ -7,12 +7,10 @@ export 'src/extensions/async.dart' show ChainAsync, FutureChecks, StreamChecks;
 export 'src/extensions/core.dart'
     show
         BoolChecks,
-        EqualityChecks,
-        HasField,
-        NullabilityChecks,
-        StringChecks,
-        TypeChecks;
+        CoreChecks,
+        NullabilityChecks;
 export 'src/extensions/function.dart' show ThrowsCheck;
 export 'src/extensions/iterable.dart' show IterableChecks;
 export 'src/extensions/map.dart' show MapChecks;
 export 'src/extensions/math.dart' show NumChecks;
+export 'src/extensions/string.dart' show StringChecks;
